Add map utility functions and update location handling
Introduced MapUtility for AoE and targeting logic, and LocationExtensions for easier location updates. Updated GameClient to use new map ID lists from GameMap. Made Location and related model properties settable. Improved consigned shop refresh logic and logging. Adjusted repository and attribute experience logic for consistency.

+++ src/Source/Distribution/DigitalWorldOnline.Game.Host/PvpServers/PvpServerBaseOperation.cs
@@ -442,7 +442,7 @@ namespace DigitalWorldOnline.GameHost
 
         // ----------------------------------------------------------------------------------------------------------------------------
 
-        public DigimonModel? GetEnemyByHandler(short mapId, int handler)
+        public DigimonModel? GetEnemyByHandler(long mapId, int handler)
         {
             return Maps.FirstOrDefault(x => x.MapId == mapId)?.ConnectedTamers
                 .Select(x => x.Partner).FirstOrDefault(x => x.GeneralHandler == handler);
